requirements.txt
# Core
streamlit==1.31.0
pandas==2.2.0
python-dotenv==1.0.1

# Autenticación
streamlit-authenticator==0.3.2
PyYAML==6.0.1
bcrypt==4.1.2

# Base de datos
openpyxl==3.1.2

# Google Sheets
gspread==5.12.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Notificaciones por Email
resend==0.8.0

# Notificaciones programadas
APScheduler==3.10.4

# WhatsApp (opcional)
twilio==8.13.0

# Manejo de imágenes
Pillow==10.2.0

# Utilidades
pytz==2024.1
